Make short URLs work on regular /experiment page
